refactor: move page components to pages directory
