+++ client/src/__tests__/App.snapshot.test.tsx
@@ -5,7 +5,7 @@ import toJson from "enzyme-to-json";
 import App, { ProtectedAppContainer } from "../App";
 import { mockUserLoggedIn, mockUserNotLoggedIn } from "../__mocks__/auth";
 import { AuthProvider } from "../components/AuthProvider";
-import { Login } from "../components/Login/Login";
+import { Login } from "../pages/Login/Login";
 import { AppSidebar, AppAdminSidebarMenu } from "../layout/AppSidebar";
 
 it("renders App correctly", async () => {
