+++ client/src/__tests__/DatasetDetails.test.tsx
@@ -7,7 +7,7 @@ import {
   within,
 } from "@testing-library/react";
 import React from "react";
-import { DatasetDetails } from "../components/DatasetDetails/DatasetDetails";
+import { DatasetDetails } from "../pages/DatasetDetails/DatasetDetails";
 import { createMemoryHistory } from "history";
 import { Router } from "react-router-dom";
 import userEvent from "@testing-library/user-event";
