+++ client/src/pages/DataEntry/DataEntryCategorySections.tsx
