+++ client/src/pages/Home/HomeSearchAutoComplete.css
@@ -1,7 +1,7 @@
 .ant-select-selection-placeholder {
-    color: #000;
+  color: #000;
 }
 
 .ant-select-item-empty::first-letter {
-    text-transform: capitalize;
+  text-transform: capitalize;
 }
