+++ client/src/pages/Home/HomeSearchAutoComplete.tsx
